+++ processor/src/test/resources/expected/com/example/repository/Arez_CompleteRepositoryExample.java
@@ -179,9 +179,4 @@ final class Arez_CompleteRepositoryExample extends CompleteRepositoryExample imp
       return super.toString();
     }
   }
-
-  @FunctionalInterface
-  interface OnDispose {
-    void onDispose(Arez_CompleteRepositoryExample entity);
-  }
 }