+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithMultipleCtors.java
@@ -152,9 +152,4 @@ final class Arez_RepositoryWithMultipleCtors extends RepositoryWithMultipleCtors
       return super.toString();
     }
   }
-
-  @FunctionalInterface
-  interface OnDispose {
-    void onDispose(Arez_RepositoryWithMultipleCtors entity);
-  }
 }