+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithProtectedConstructor.java
@@ -128,9 +128,4 @@ final class Arez_RepositoryWithProtectedConstructor extends RepositoryWithProtec
       return super.toString();
     }
   }
-
-  @FunctionalInterface
-  interface OnDispose {
-    void onDispose(Arez_RepositoryWithProtectedConstructor entity);
-  }
 }