+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithSingleton.java
@@ -128,9 +128,4 @@ final class Arez_RepositoryWithSingleton extends RepositoryWithSingleton impleme
       return super.toString();
     }
   }
-
-  @FunctionalInterface
-  interface OnDispose {
-    void onDispose(Arez_RepositoryWithSingleton entity);
-  }
 }