+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Unreleased
 
+#### Changed
+* **\[processor\]** Remove the unused nested class `OnDispose` that was added to enhanced component classes.
+
 ### [v0.48](https://github.com/arez/arez/tree/v0.48) (2018-02-02)
 [Full Changelog](https://github.com/arez/arez/compare/v0.47...v0.48)